refactor(trawling): Consolidate overlay classes into unified TrawlingOverlay
- Delete separate NetCapacityOverlay and NetDepthTimerOverlay classes
- Create new TrawlingOverlay that combines both net capacity and depth timer displays
- Consolidate overlay rendering logic into single panel with conditional sections
- Update SailingModule to register unified overlay instead of separate overlays
- Improve maintainability by reducing duplicate overlay infrastructure code
