+++ src/main/java/com/duckblade/osrs/sailing/module/SailingModule.java
@@ -36,12 +36,11 @@ import com.duckblade.osrs.sailing.features.oceanencounters.LostShipment;
 import com.duckblade.osrs.sailing.features.oceanencounters.MysteriousGlow;
 import com.duckblade.osrs.sailing.features.oceanencounters.OceanMan;
 import com.duckblade.osrs.sailing.features.salvaging.SalvagingHighlight;
-import com.duckblade.osrs.sailing.features.trawling.NetCapacityOverlay;
 import com.duckblade.osrs.sailing.features.trawling.NetCapacityTracker;
 import com.duckblade.osrs.sailing.features.trawling.NetDepthButtonHighlighter;
 import com.duckblade.osrs.sailing.features.trawling.NetDepthTimer;
 import com.duckblade.osrs.sailing.features.trawling.NetDepthTracker;
-import com.duckblade.osrs.sailing.features.trawling.NetDepthTimerOverlay;
+import com.duckblade.osrs.sailing.features.trawling.TrawlingOverlay;
 import com.duckblade.osrs.sailing.features.trawling.ShoalDepthTracker;
 import com.duckblade.osrs.sailing.features.trawling.ShoalOverlay;
 import com.duckblade.osrs.sailing.features.trawling.ShoalPathTrackerOverlay;
@@ -99,12 +98,11 @@ public class SailingModule extends AbstractModule
 		CrystalExtractorHighlight crystalExtractorHighlight,
 		MermaidTaskSolver mermaidTaskSolver,
 		MysteriousGlow mysteriousGlow,
-		NetCapacityOverlay netCapacityOverlay,
 		NetCapacityTracker netCapacityTracker,
 		NetDepthButtonHighlighter netDepthButtonHighlighter,
 		NetDepthTimer netDepthTimer,
 		NetDepthTracker netDepthTracker,
-		NetDepthTimerOverlay netDepthTimerOverlay,
+		TrawlingOverlay trawlingOverlay,
 		OceanMan oceanMan,
 		ShoalDepthTracker shoalDepthTracker,
 		PrioritizeCargoHold prioritizeCargoHold,
@@ -153,11 +151,10 @@ public class SailingModule extends AbstractModule
 			.add(crystalExtractorHighlight)
 			.add(mermaidTaskSolver)
 			.add(mysteriousGlow)
-			.add(netCapacityOverlay)
 			.add(netCapacityTracker)
 			.add(netDepthButtonHighlighter)
 			.add(netDepthTimer)
-			.add(netDepthTimerOverlay)
+			.add(trawlingOverlay)
 			.add(netDepthTracker)
 			.add(navigationOverlay)
 			.add(oceanMan)
